Remove elements and kotlinmetadata from property, constructor, type
