+++ kotlin/codegen/src/main/java/com/squareup/moshi/kotlin/codegen/AdapterGenerator.kt
@@ -47,14 +47,14 @@ internal class AdapterGenerator(
     private val propertyList: List<PropertyGenerator>
 ) {
   private val nonTransientProperties = propertyList.filterNot { it.isTransient }
-  private val className = target.name
+  private val className = target.typeName.rawType()
   private val visibility = target.visibility
   private val typeVariables = target.typeVariables
 
   private val nameAllocator = NameAllocator()
   private val adapterName = "${className.simpleNames.joinToString(separator = "_")}JsonAdapter"
-  private val originalElement = target.element
-  private val originalTypeName = target.element.asType().asTypeName()
+  private val originalElement = target.element // TODO fix after rebase
+  private val originalTypeName = target.typeName
 
   private val moshiParam = ParameterSpec.builder(
       nameAllocator.newName("moshi"),
